Fix one-to-many backref deleting
Previously, when deleting a service, SQLAlchemy would attempt to
modify the corresponding messages and subscriptions to a service
because it does not assume cascade deletion. Instead it attempted
to set the foreign keys in subscription and services tables to NULL
which is prevented by the SQL consistency rules.

This patch enables cascade deletion on these backrefs.

Now all tests pass running under python 3